+++ src/components/Header/Header.tsx
@@ -1,55 +1,13 @@
 import { LoginOutlined } from '@mui/icons-material';
 import { AppBar, IconButton, Toolbar, Link, Typography } from '@mui/material';
 import { useState } from 'react';
-import { styled } from '@mui/material/styles';
-import Switch from '@mui/material/Switch';
 import Stack from '@mui/material/Stack';
+import LangSwitch from '../../styles/MUI/langSwitch';
 
 const LANG = {
   rus: 'RUS',
   eng: 'ENG',
 };
-
-const LangSwitch = styled(Switch)(({ theme }) => ({
-  width: 28,
-  height: 16,
-  padding: 0,
-  display: 'flex',
-  '&:active': {
-    '& .MuiSwitch-thumb': {
-      width: 15,
-    },
-    '& .MuiSwitch-switchBase.Mui-checked': {
-      transform: 'translateX(9px)',
-    },
-  },
-  '& .MuiSwitch-switchBase': {
-    padding: 2,
-    '&.Mui-checked': {
-      transform: 'translateX(12px)',
-      color: '#fff',
-      '& + .MuiSwitch-track': {
-        opacity: 1,
-        backgroundColor: '#1890ff',
-      },
-    },
-  },
-  '& .MuiSwitch-thumb': {
-    boxShadow: '0 2px 4px 0 rgb(0 35 11 / 20%)',
-    width: 12,
-    height: 12,
-    borderRadius: 6,
-    transition: theme.transitions.create(['width'], {
-      duration: 200,
-    }),
-  },
-  '& .MuiSwitch-track': {
-    borderRadius: 16 / 2,
-    opacity: 1,
-    backgroundColor: '#1890ff',
-    boxSizing: 'border-box',
-  },
-}));
 
 function Header() {
   const [lang, setLanguage] = useState(LANG.eng);
