+++ src/main.tsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import ReactDOM from 'react-dom/client';
-import { ThemeProvider, createTheme } from '@mui/material';
+import { ThemeProvider } from '@mui/material';
 import App from './App';
 import './index.css';
 import '@fontsource/roboto/300.css';
@@ -8,38 +8,7 @@ import '@fontsource/roboto/400.css';
 import '@fontsource/roboto/500.css';
 import '@fontsource/roboto/700.css';
 
-const theme = createTheme({
-  palette: {
-    primary: {
-      main: '#673ab7',
-    },
-    secondary: {
-      main: '#d500f9',
-    },
-  },
-  components: {
-    MuiLink: {
-      styleOverrides: {
-        root: {
-          textDecoration: 'none',
-          ':hover': {
-            color: '#d500f9',
-          },
-        },
-      },
-    },
-    MuiIconButton: {
-      styleOverrides: {
-        root: {
-          textDecoration: 'none',
-          ':hover': {
-            color: '#d500f9',
-          },
-        },
-      },
-    },
-  },
-});
+import theme from './styles/theme';
 
 ReactDOM.createRoot(document.getElementById('root')!).render(
   <React.StrictMode>
