+++ src/setupTests.ts
@@ -1,7 +1,7 @@
 /* eslint-disable import/no-extraneous-dependencies */
-import { afterEach } from "vitest";
-import { cleanup } from "@testing-library/react";
-import "@testing-library/jest-dom";
+import { afterEach } from 'vitest';
+import { cleanup } from '@testing-library/react';
+import '@testing-library/jest-dom';
 
 afterEach(() => {
   cleanup();
